🔥Longs (#17)
* Kill longs

* Revert this

* remove unnecessary cast

* Fix the build

+++ lib/src/armMain/kotlin/com/felipecsl/knes/platform.kt
@@ -6,10 +6,10 @@ import kotlinx.cinterop.ptr
 import platform.posix.gettimeofday
 import platform.posix.timeval
 
-actual fun currentTimeMs(): Long {
+actual fun currentTimeMs(): Double {
   memScoped {
     val now = alloc<timeval>()
     gettimeofday(now.ptr, null)
-    return (now.tv_sec.toLong() * 1000) + (now.tv_usec.toLong() / 1000)
+    return (now.tv_sec * 1000.0) + (now.tv_usec / 1000.0)
   }
 }
